+++ test/conftest.py
@@ -10,8 +10,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from __future__ import unicode_literals
-from __future__ import print_function
 
 import itertools
 import os
@@ -19,9 +17,9 @@ import subprocess
 import sys
 import threading
 import time
+import urllib.parse
 
 import pytest
-from six.moves import urllib
 
 import testinfra
 from testinfra.backend.base import BaseBackend