+++ testinfra/backend/openshift.py
@@ -14,32 +14,32 @@
 from testinfra.backend import base
 
 
-class KubectlBackend(base.BaseBackend):
-    NAME = "kubectl"
+class OpenShiftBackend(base.BaseBackend):
+    NAME = "openshift"
 
     def __init__(self, name, *args, **kwargs):
         self.name = name
         self.container = kwargs.get('container')
         self.namespace = kwargs.get('namespace')
         self.kubeconfig = kwargs.get('kubeconfig')
-        super(KubectlBackend, self).__init__(self.name, *args, **kwargs)
+        super(OpenShiftBackend, self).__init__(self.name, *args, **kwargs)
 
     def run(self, command, *args, **kwargs):
         cmd = self.get_command(command, *args)
-        # `kubectl exec` does not support specifying the user to run as.
+        # `oc exec` does not support specifying the user to run as.
         # See https://github.com/kubernetes/kubernetes/issues/30656
-        kcmd = 'kubectl '
-        kcmd_args = []
+        oscmd = 'oc '
+        oscmd_args = []
         if self.kubeconfig is not None:
-            kcmd += '--kubeconfig="%s" '
-            kcmd_args.append(self.kubeconfig)
+            oscmd += '--kubeconfig="%s" '
+            oscmd_args.append(self.kubeconfig)
         if self.namespace is not None:
-            kcmd += '-n %s '
-            kcmd_args.append(self.namespace)
+            oscmd += '-n %s '
+            oscmd_args.append(self.namespace)
         if self.container is not None:
-            kcmd += '-c %s '
-            kcmd_args.append(self.container)
-        kcmd += 'exec %s -- /bin/sh -c %s'
-        kcmd_args.extend([self.name, cmd])
-        out = self.run_local(kcmd, *kcmd_args)
+            oscmd += '-c %s '
+            oscmd_args.append(self.container)
+        oscmd += 'exec %s -- /bin/sh -c %s'
+        oscmd_args.extend([self.name, cmd])
+        out = self.run_local(oscmd, *oscmd_args)
         return out