+++ testinfra/modules/__init__.py
@@ -11,8 +11,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import unicode_literals
-
 import importlib
 
 modules = {
@@ -20,6 +18,7 @@ modules = {
     'ansible': 'ansible:Ansible',
     'command': 'command:Command',
     'docker': 'docker:Docker',
+    'podman': 'podman:Podman',
     'environment': 'environment:Environment',
     'file': 'file:File',
     'group': 'group:Group',