+++ testinfra/modules/ansible.py
@@ -11,9 +11,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import
-from __future__ import unicode_literals
-
 import functools
 import pprint
 
@@ -64,6 +61,20 @@ class Ansible(InstanceModule):
     <https://docs.ansible.com/ansible/user_guide/become.html#id1>`_ is
     `False` by default. You can enable it with `become=True`.
 
+    Ansible arguments that are not related to the Ansible inventory or
+    connection (both managed by testinfra) are also accepted through keyword
+    arguments:
+
+        - ``become_method`` *str* sudo, su, doas, etc.
+        - ``become_user`` *str* become this user.
+        - ``diff`` *bool*: when changing (small) files and templates, show the
+          differences in those files.
+        - ``extra_vars`` *dict* serialized to a JSON string, passed to
+          Ansible.
+        - ``one_line`` *bool*: condense output.
+        - ``user`` *str* connect as this user.
+        - ``verbose`` *int* level of verbosity
+
     >>> host.ansible("apt", "name=nginx state=present")["changed"]
     False
     >>> host.ansible("apt", "name=nginx state=present", become=True)["changed"]
@@ -74,6 +85,21 @@ class Ansible(InstanceModule):
     'jessie'
     >>> host.ansible("file", "path=/etc/passwd")["mode"]
     '0640'
+    >>> host.ansible(
+    ... "command",
+    ... "id --user --name",
+    ... check=False,
+    ... become=True,
+    ... become_user="http",
+    ... )["stdout"]
+    'http'
+    >>> host.ansible(
+    ... "apt",
+    ... "name={{ packages }}",
+    ... check=False,
+    ... extra_vars={"packages": ["neovim", "vim"]},
+    ... )
+    # Installs neovim and vim.
 
     """
     # pylint: disable=self-assigning-variable