+++ testinfra/utils/ansible_runner.py
@@ -11,20 +11,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import unicode_literals
-from __future__ import absolute_import
-
+import configparser
 import fnmatch
+import ipaddress
 import json
 import os
+import tempfile
 
-from six.moves import configparser
 
 import testinfra
 from testinfra.utils import cached_property
 from testinfra.utils import check_ip_address
 from testinfra.utils import TemporaryDirectory
 import yaml
+
 
 __all__ = ['AnsibleRunner']
 
@@ -80,6 +80,7 @@ def get_ansible_host(config, inventory, host, ssh_config=None,
     }.get(connection, connection)
     testinfra_host = hostvars.get('ansible_host', host)
     user = hostvars.get('ansible_user')
+    password = hostvars.get('ansible_ssh_pass')
     port = hostvars.get('ansible_port')
     kwargs = {}
     if hostvars.get('ansible_become', False):
@@ -103,9 +104,18 @@ def get_ansible_host(config, inventory, host, ssh_config=None,
     ).strip()
 
     spec = '{}://'.format(connection)
-    if user:
+
+    # Fallback to user:pasword auth when identity file is not used
+    if user and password and not kwargs.get('ssh_identity_file'):
+        spec += '{}:{}@'.format(user, password)
+    elif user:
         spec += '{}@'.format(user)
-    if check_ip_address(testinfra_host) == 6:
+
+    try:
+        version = ipaddress.ip_address(testinfra_host).version
+    except ValueError:
+        version = None
+    if version == 6:
         spec += '[' + testinfra_host + ']'
     else:
         spec += testinfra_host
@@ -128,6 +138,43 @@ def is_empty_inventory(inventory):
 
 class AnsibleRunner(object):
     _runners = {}
+    _known_options = {
+        # Boolean arguments.
+        "become": {
+            "cli": "--become",
+            "type": "boolean",
+        },
+        "check": {
+            "cli": "--check",
+            "type": "boolean",
+        },
+        "diff": {
+            "cli": "--diff",
+            "type": "boolean",
+        },
+        "one_line": {
+            "cli": "--one-line",
+            "type": "boolean",
+        },
+        # String arguments.
+        "become_method": {
+            "cli": "--become-method",
+            "type": "string",
+        },
+        "become_user": {
+            "cli": "--become-user",
+            "type": "string",
+        },
+        "user": {
+            "cli": "--user",
+            "type": "string",
+        },
+        # Arguments serialized as JSON.
+        "extra_vars": {
+            "cli": "--extra-vars",
+            "type": "json",
+        },
+    }
 
     def __init__(self, inventory_file=None, ansible_playbook=None):
         self.inventory_file = inventory_file
@@ -201,8 +248,35 @@ class AnsibleRunner(object):
                 self.ansible_config, self.inventory, host, **kwargs)
             return self._host_cache[host]
 
-    def run_module(self, host, module_name, module_args, become=False,
-                   check=True, **kwargs):
+    def options_to_cli(self, options):
+        verbose = options.pop("verbose", 0)
+
+        args = {"become": False, "check": True}
+        args.update(options)
+
+        cli = []
+        cli_args = []
+        if verbose:
+            cli.append('-' + "v" * verbose)
+        for arg_name, value in args.items():
+            option = self._known_options[arg_name]
+            opt_cli = option["cli"]
+            opt_type = option["type"]
+            if opt_type == "boolean":
+                if value:
+                    cli.append(opt_cli)
+            elif opt_type == "string":
+                cli.append(opt_cli + " %s")
+                cli_args.append(value)
+            elif opt_type == "json":
+                cli.append(opt_cli + " %s")
+                value_json = json.dumps(value)
+                cli_args.append(value_json)
+            else:
+                raise TypeError("Unsupported argument type '%s'." % opt_type)
+        return " ".join(cli), cli_args
+
+    def run_module(self, host, module_name, module_args, **options):
         cmd, args = 'ansible --tree %s', [None]
         if self.inventory_file:
             cmd += ' -i %s'
@@ -212,13 +286,13 @@ class AnsibleRunner(object):
         if module_args:
             cmd += ' --args %s'
             args += [module_args]
-        if become:
-            cmd += ' --become'
-        if check:
-            cmd += ' --check'
+        options_cli, options_args = self.options_to_cli(options)
+        if options_cli:
+            cmd += ' ' + options_cli
+            args.extend(options_args)
         cmd += ' %s'
         args += [host]
-        with TemporaryDirectory() as d:
+        with tempfile.TemporaryDirectory() as d:
             args[0] = d
             out = local.run_expect([0, 2, 8], cmd, *args)
             files = os.listdir(d)